Add files via upload
I'm including each version of Stab-acus from the development period, so I apologize for the confusing name structure. The build is comprised of stabacus.cpp (runs the game), hosttext.h (quirky text to be displayed at key moments), and stabacus-game.cpp & stabacus-game.h (define functions needed to play the game, such as Stabbing and Moving). 
stabacusv1.out is the current build that should be able to run independently; it includes all of these files.