Add files via upload
Encountered a late-game bug in which one player running out of balls caused an infinite loop; resolved it and added new dialogue options for this situation.
Some quality-of-life updates. 
stab-acusv2.exe is now the most updated version of the game! Still looking into making a separate, stand-alone executable (one that doesn't need an IDE).